fix: more warnings
