+++ src/lock_file/resolve/pypi.rs
@@ -1,10 +1,11 @@
 use super::pypi_editables::build_editables;
 use crate::consts::PROJECT_MANIFEST;
+use crate::lock_file::resolve::resolver_provider::CondaResolverProvider;
 use crate::project::manifest::pypi_options::PypiOptions;
 use crate::project::manifest::python::RequirementOrEditable;
 use crate::uv_reporter::{UvReporter, UvReporterOptions};
-use std::collections::{BTreeMap, HashMap};
-use std::future::{ready, Future};
+use std::collections::HashMap;
+
 use std::iter::once;
 
 use crate::lock_file::{
@@ -17,12 +18,11 @@ use crate::{
     project::manifest::{PyPiRequirement, SystemRequirements},
 };
 
+use distribution_types::FileLocation;
 use distribution_types::{
-    BuiltDist, DirectUrlSourceDist, Dist, FlatIndexLocation, HashPolicy, IndexLocations, IndexUrl,
-    LocalEditable, Name, PrioritizedDist, Resolution, ResolvedDist, SourceDist,
+    BuiltDist, Dist, FlatIndexLocation, HashPolicy, IndexUrl, LocalEditable, Name, Resolution,
+    ResolvedDist, SourceDist,
 };
-use distribution_types::{FileLocation, SourceDistCompatibility};
-use futures::FutureExt;
 use indexmap::{IndexMap, IndexSet};
 use indicatif::ProgressBar;
 use install_wheel_rs::linker::LinkMode;
@@ -44,15 +44,14 @@ use uv_configuration::{ConfigSettings, Constraints, NoBuild, Overrides, SetupPyS
 use url::Url;
 use uv_client::{Connectivity, FlatIndexClient, RegistryClient, RegistryClientBuilder};
 use uv_dispatch::BuildDispatch;
-use uv_distribution::{ArchiveMetadata, DistributionDatabase, Reporter};
+use uv_distribution::DistributionDatabase;
 use uv_interpreter::Interpreter;
 use uv_normalize::PackageName;
 use uv_resolver::{
-    AllowedYanks, DefaultResolverProvider, FlatIndex, InMemoryIndex, Manifest, MetadataResponse,
-    Options, Preference, PythonRequirement, Resolver, ResolverProvider, VersionMap,
-    VersionsResponse, WheelMetadataResult,
+    AllowedYanks, DefaultResolverProvider, FlatIndex, InMemoryIndex, Manifest, Options, Preference,
+    PythonRequirement, Resolver,
 };
-use uv_types::{BuildContext, EmptyInstalledPackages, HashStrategy, InFlight};
+use uv_types::{BuildContext, EmptyInstalledPackages};
 
 fn parse_hashes_from_hash_vec(hashes: &Vec<HashDigest>) -> Option<PackageHashes> {
     let mut sha256 = None;
@@ -84,89 +83,6 @@ fn parse_hashes_from_hash_vec(hashes: &Vec<HashDigest>) -> Option<PackageHashes>
             parse_digest_from_hex::<Sha256>(&sha256).expect("invalid sha256"),
         )),
         (None, None) => None,
-    }
-}
-
-struct CondaResolverProvider<'a, Context: BuildContext + Send + Sync> {
-    fallback: DefaultResolverProvider<'a, Context>,
-    conda_python_identifiers: &'a HashMap<PackageName, (RepoDataRecord, PypiPackageIdentifier)>,
-}
-
-impl<'a, Context: BuildContext + Send + Sync> ResolverProvider
-    for CondaResolverProvider<'a, Context>
-{
-    fn get_package_versions<'io>(
-        &'io self,
-        package_name: &'io PackageName,
-    ) -> impl Future<Output = uv_resolver::PackageVersionsResult> + Send + 'io {
-        if let Some((repodata_record, identifier)) = self.conda_python_identifiers.get(package_name)
-        {
-            // If we encounter a package that was installed by conda we simply return a single
-            // available version in the form of a source distribution with the URL of the
-            // conda package.
-            //
-            // Obviously this is not a valid source distribution but it easies debugging.
-            let dist = Dist::Source(SourceDist::DirectUrl(DirectUrlSourceDist {
-                name: identifier.name.as_normalized().clone(),
-                url: VerbatimUrl::unknown(repodata_record.url.clone()),
-            }));
-
-            let prioritized_dist = PrioritizedDist::from_source(
-                dist,
-                Vec::new(),
-                SourceDistCompatibility::Compatible(distribution_types::Hash::Matched),
-            );
-
-            return ready(Ok(VersionsResponse::Found(vec![VersionMap::from(
-                BTreeMap::from_iter([(identifier.version.clone(), prioritized_dist)]),
-            )])))
-            .right_future();
-        }
-
-        // Otherwise use the default implementation
-        self.fallback
-            .get_package_versions(package_name)
-            .left_future()
-    }
-
-    fn get_or_build_wheel_metadata<'io>(
-        &'io self,
-        dist: &'io Dist,
-    ) -> impl Future<Output = WheelMetadataResult> + Send + 'io {
-        if let Dist::Source(SourceDist::DirectUrl(DirectUrlSourceDist { name, .. })) = dist {
-            if let Some((_, iden)) = self.conda_python_identifiers.get(name) {
-                // If this is a Source dist and the package is actually installed by conda we
-                // create fake metadata with no dependencies. We assume that all conda installed
-                // packages are properly installed including its dependencies.
-                return ready(Ok(MetadataResponse::Found(ArchiveMetadata {
-                    metadata: Metadata23 {
-                        name: iden.name.as_normalized().clone(),
-                        version: iden.version.clone(),
-                        requires_dist: vec![],
-                        requires_python: None,
-                        provides_extras: iden.extras.iter().cloned().collect(),
-                    },
-                    hashes: vec![],
-                })))
-                .left_future();
-            }
-        }
-
-        // Otherwise just call the default implementation
-        self.fallback
-            .get_or_build_wheel_metadata(dist)
-            .right_future()
-    }
-
-    fn index_locations(&self) -> &IndexLocations {
-        self.fallback.index_locations()
-    }
-
-    fn with_reporter(self, reporter: impl Reporter + 'static) -> Self {
-        Self {
-            fallback: self.fallback.with_reporter(reporter),
-            ..self
-        }
     }
 }
 
@@ -517,7 +433,7 @@ pub async fn resolve_pypi(
     .await
 }
 
-/// Create a vector of locked pacakges from a resolution
+/// Create a vector of locked packages from a resolution
 async fn lock_pypi_packages<'a>(
     conda_python_packages: CondaPythonPackages,
     build_dispatch: &BuildDispatch<'a>,
